% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbyEachClass.R
\name{fbyEachClass}
\alias{fbyEachClass}
\title{Context for rows and columns}
\usage{
fbyEachClass(r_, f_, d_)
}
\arguments{
\item{r_}{the row}

\item{f_}{the feature}

\item{d_}{reference to the dataframe}
}
\value{
a normed datarfame`
}
\description{
Context for rows and columns
}
