% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qGEL.R
\name{qgel}
\alias{qgel}
\title{qgel}
\arguments{
\item{source.data_}{the data to embed}

\item{k}{number of embedded features, default k = 10}

\item{class_var}{string, the name of class variable}

\item{learning_method}{string, 'supervised' otherwise will be unsupervised}
}
\description{
qgel
}
